Big improvement for the Colliders
Now instead of four subclasses, one for each direction that sonic can move towards, there is only one classe, called PlayerCollider.
I also modified the Collider superclass in order to reorganize that part of the physics engine, so that all the type of colliders will be extensions of Collider.java
Now Collider.java no longer extends JPanel, but has a method to update his position to Sonic's. This was made to ease the detection of collisions with all the others element of the JFrame.
Lastly, i created EnemyCollider and RingCollider, that will be inserted as attributes in every instance of Ring or Enemy