Enemy phisics created
Now the EnemyColliders can detect walls too, not just the player. This let the other classes know what is the limit for the movement of the Enemy. I Worked on the EnemyPhisics class, which has the purpose to make the enemy chase Sonic as long as possible. and determine what action is the enemy doing and which direction he's facing